+++ lib/protobuf/rpc/middleware.rb
@@ -4,6 +4,7 @@ require 'protobuf/rpc/middleware/exception_handler'
 require 'protobuf/rpc/middleware/logger'
 require 'protobuf/rpc/middleware/request_decoder'
 require 'protobuf/rpc/middleware/response_encoder'
+require 'protobuf/rpc/middleware/statsd'
 require 'protobuf/rpc/middleware/runner'
 
 module Protobuf
@@ -16,6 +17,8 @@ module Protobuf
     middleware
   end
 
+  # Statsd comes first so it gets full timing and access to exceptions
+  Rpc.middleware.use(Rpc::Middleware::Statsd)
   Rpc.middleware.use(Rpc::Middleware::ExceptionHandler)
   Rpc.middleware.use(Rpc::Middleware::RequestDecoder)
   Rpc.middleware.use(Rpc::Middleware::Logger)